+++ lib/worker.js
@@ -3,7 +3,7 @@ const db = require('@asl/schema');
 const rops = require('./exporters/rops');
 const taskMetrics = require('./exporters/task-metrics');
 const Logger = require('./utils/logger');
-const S3 = require('./clients/s-three');
+const s3Upload = require('./clients/s3-upload');
 
 module.exports = settings => {
   const logger = Logger(settings);
@@ -11,7 +11,7 @@ module.exports = settings => {
 
   settings.models = db(settings.db);
 
-  settings.s3Upload = S3(settings.s3);
+  settings.s3Upload = s3Upload(settings.s3);
 
   const exporters = {
     rops: rops(settings),
